requirements.txt
opencv-python-headless="*"
deepmerge>=1.0.1
dotmap>=1.3.30
jsonschema>=4.16.0
matplotlib>=3.5.3
numpy>=1.23.2
pandas>=1.4.3
rich>=12.6.0
screeninfo>=0.8.1
flask="*"
pymupdf="*"
pillow="*"
flask-json="*"